Add functionality to broadcast concurrently to multiple clients
The current implementation uses a queue to keep track of all the
payloads constructed by the thread corresponding to 'webcamFeed'.
The consumer (thread corresponding to 'handleConnection') checks the
queue if it contains the payload, dequeues and sends over the socket if
found. This greatly simplifies the logic for keeping track of the next
payload which is to be served.

An ideal implementation will main a list of payloads which are to be
served, and each consumer should keep track of the payload which it
served last and look out for the next payload (if ready) in the list.

Each payload in the list should have a lifetime, after which it is
considered stale and automatically removed. This will ensure a constant
size of the list over a period of time.

It should be noted that the current setup is tested to work with 2
clients. However, this should not be the case unless the consumers
corresponding to both the clients are able to perform a successful
dequeue operation -- the operations should (ideally) be made atomic.

Fixes #1
Fixes #6