Improve server performance
The number of CPU bursts for the webcam thread were huge as the time
required to collect 'chunk_size' worth of frames is not negligible. Thus
the webcam thread ended up yielding the CPU as soon as it's burst
started almost everytime.

This is fixed by making the consumer thread sleep for 0.1 seconds which
gives the webcam thread enough time to fill up the queue.

This change improved the performance by reducing the number of CPU
bursts for webcamFeed to 98 from 3634768, corresponding to a fixed
number of 126 CPU bursts for handleConnection.