Update sleep durations with more accurate values
TODO
----
There seems to be some error while delivering and retrieving a payload.
For instance, this is the data from one of the test runs -

Server statistics
-----------------
No. of CPU yields -
  * handleConnection: 14
  * webcamFeed: 4
Frames per payload: 5
Payloads delivered: 38

Client statistics
-----------------
Frames displayed: 37

The number of displayed frames on the client side doesn't match the
total number of frames delivered by the server.