server: impose a strict lower bound on the size of the payload buffer
At any given instant, the gap between a reader and the writer can be
atmost lag_threshold. This is also the lower bound on the size of the
payload buffer. Also, the maximum allowed lag is relaxed to avoid
frequent jerks in the video.
This greatly reduces the memory footprint.